Public

Public documentation

Documentation for public SeisProcessing.jl public interface

Public interface

SeisReconstruction.SeisPOCS — Function.

SeisPOCS(in;<keyword arguments>)

Projection Onto Convex Sets interpolation of seismic records.

Arguments

in: input data that can have up to 5 dimensions. Time is in the first dimension.

p=1. : exponent for thresholding (1 is equivalent to soft thres. high number is equivalent to hard thresholding)

alpha=1 : add-back ratio for imputation step. Use 1 for noise free data, and < 1 for denoising of original traces.

dt=0.001 : sampling rate along the time axis (in seconds)

fmax=99999. : maximum temporal frequency to process.

padt=2 : padding in the time axis, first dimension.

padx=1 : padding in the spatial axes. (Dim 2 to 5).

Niter=100 : number of iterations

alpha=1

source

SeisReconstruction.CalculateSampling — Function.

CalculateSampling(in)

Calculate the sampling operator of an n-dimension input. The output has the same size as the input.

source

SeisReconstruction.ConjugateGradients — Function.

ConjugateGradients(d,operators,parameters;<keyword arguments>)

Conjugate Gradients following Algorithm 2 from Scales, 1987. The user provides an array of linear operators. Verify that linear operator(s) pass the dot product. See also: DotTest

Arguments

Niter=10 : Number of iterations

mu=0

tol=1.0e-15

source

SeisReconstruction.DotTest — Function.

DotTest(m_rand,d_rand,operators,parameters)

Dot product test for a vector of linear operators See also: ConjugateGradients

source

SeisReconstruction.IRLS — Function.

IRLS(d,operators,parameters;<keyword arguments>)

Non-quadratic regularization with Iteratively Reweighted Least Squares (IRLS).

Arguments

Niter_external=3

'Niter_internal=10'

mu=0

source